2017.2.21 new

+++ comedu/comedu_calendar/views.py
@@ -3,9 +3,8 @@ from django.views.generic import ListView
 from .models import CalendarEvent
 from .forms import CalendarForm
 from django.http import HttpResponse
-from django.template import RequestContext
-from django.contrib import messages
-
+from django.core.paginator import Paginator
+from django.conf import settings
 
 # Create your views here.
 
@@ -20,7 +19,7 @@ class CalendarEventLV(ListView):
 
 def cal_detail(request, pk):
     if not request.user.is_authenticated:
-        return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
+        return redirect('/calendar/')
     else:
         queryset2 = get_object_or_404(CalendarEvent, pk=pk)
         return render(request, 'comedu_calendar/calendar_detail.html', {'calendar':queryset2})
@@ -29,37 +28,37 @@ def cal_detail(request, pk):
 
 def calendar_new(request):
     if not request.user.is_authenticated:
-        return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
-
+        return redirect('/calendar/')
     else:
         if request.method == "POST":
-            form = CalendarForm(request.POST)##if method = POST save all content in site which sended request in form
-            if form.is_valid():
-                calendar = form.save(commit=False)##don't save directly, because should save after saving author info
-                #calendar.save()
-                return redirect('/calendar/')
+                form = CalendarForm(request.POST)##if method = POST save all content in site which sended request in form
+                if form.is_valid():
+                    calendar = form.save(commit=False)
+                    calendar.save()
+                    return redirect('/calendar/')
         else:
             form = CalendarForm()##????
 
-    return render(request, 'comedu_calendar/calendar_new.html', {'form': form})##why give 'form':form???
+        return render(request, 'comedu_calendar/calendar_new.html', {'form': form})##why give 'form':form???
 
 
 
 def calendar_edit(request, pk=None, template_name='comedu_calendar/calendar_edit.html'):##pk값이 없으면 None
-    if pk:
-        calendar = get_object_or_404(CalendarEvent,id=pk)#저 모델에서 받은 pk값을 가지고 있는 모델 가져오기
+    if not request.user.is_authenticated:
+        return redirect('/calendar/')
     else:
-        return HttpResponseForbidden()
+        if pk:
+            calendar = get_object_or_404(CalendarEvent,id=pk)#저 모델에서 받은 pk값을 가지고 있는 모델 가져오기
 
-    form = CalendarForm(request.POST or None, instance = calendar)##CalendarForm 쓰는데 calendar에 저장되어 있는 값을 디폴트로 설정
-    if request.POST and form.is_valid():##제출 받으면
-        form.save()
-        return redirect('/calendar/%s' %(pk))
+        form = CalendarForm(request.POST or None, instance = calendar)##CalendarForm 쓰는데 calendar에 저장되어 있는 값을 디폴트로 설정
 
-    return render(request, template_name, {
-        'form': form
-    })##실행 순서 : 버튼 누르면 url로 이동함과 동시에 pk값 전달 -> edit 뷰 실행 -> pk값을 가진 오브젝트를 CalendarEvent모델에서 찾아서 calendar에 저장 ->form에
-      ##원래 글 저장 -> post안 눌렀으니까 맨 마지막 줄 실행 -> form이 보임 -> 사용자가 입력 -> 제출 -> POST->if POST 밑에 실행
+        if request.POST and form.is_valid():##제출 받으면
+            form.save()
+            return redirect('/calendar/%s' %(pk))
+        if request.user == calendar.author:
+            return render(request, template_name, {
+                'form': form
+            })
 
 
 
@@ -67,25 +66,72 @@ def calendar_search(request):
     if request.GET['category'] == 'ti':
         if request.GET['q']:
             q = request.GET['q']
-            calendars = CalendarEvent.objects.filter(title__icontains = q)
-            return render_to_response('comedu_calendar/calendar_search.html', {'calendars':calendars, 'query':q})
+            search_results = CalendarEvent.objects.filter(title__icontains = q)
+            paginator = Paginator(search_results, 10)
+            page = request.GET.get('page', '1')
+            try:
+                results = paginator.page(page)
+            except PageNotAnInteger:
+                results = paginator.page(1)
+            except EmptyPage:
+                results = paginator.page(paginator.num_pages)
+            return render_to_response('comedu_calendar/calendar_search.html', {'calendars':results, 'query':q})
         else:
-            return render_to_response('calendar_search.html',{'error':True})
+            return render_to_response('comedu_calendar/calendar_search.html')
 
     if request.GET['category'] == 'co':
         if request.GET['q']:
            q = request.GET['q']
-           calendars = CalendarEvent.objects.filter(context__icontains = q)
-           return render_to_response('comedu_calendar/calendar_search.html', {'calendars':calendars, 'query':q})
+           search_results = CalendarEvent.objects.filter(context__icontains = q)
+           paginator = Paginator(search_results, 10)
+           page = request.GET.get('page', '1')
+           try:
+               results = paginator.page(page)
+           except PageNotAnInteger:
+               results = paginator.page(1)
+           except EmptyPage:
+               results = paginator.page(paginator.num_pages)
+           return render_to_response('comedu_calendar/calendar_search.html', {'calendars':results, 'query':q})
         else:
-           return render_to_response('calendar_search.html',{'error':True})
+           return render_to_response('comedu_calendar/calendar_search.html')
+
+    if request.GET['category'] == 'au':
+        if request.GET['q']:
+           q = request.GET['q']
+           search_results = CalendarEvent.objects.filter(author__username__contains = q)
+           paginator = Paginator(search_results, 10)
+           page = request.GET.get('page', '1')
+           try:
+               results = paginator.page(page)
+           except PageNotAnInteger:
+               results = paginator.page(1)
+           except EmptyPage:
+               results = paginator.page(paginator.num_pages)
+           return render_to_response('comedu_calendar/calendar_search.html', {'calendars':results, 'query':q})
+        else:
+           return render_to_response('comedu_calendar/calendar_search.html')
+
+
+def calendar_delete_popup(request, pk):
+    if not request.user.is_authenticated:
+        return redirect('/calendar/')
+    else:
+        return render_to_response('comedu_calendar/calendar_delete.html',{'pk':pk})
 
 
 
 def calendar_delete(request, pk):
-    CalendarEvent(pk=pk).delete()
-    return redirect('/calendar/')
 
+    if request.GET['answer']=="YES":
+        CalendarEvent(pk=pk).delete()
+        return redirect('/calendar/')
+    else:
+        return redirect('/calendar/%s' %(pk))
+
+
+def calendar_show(request):
+    calendars = CalendarEvent.objects.all()
+    return render(request, 'comedu_calendar/calendar.html', {'calendars':calendars})
 
 
 def calendar_success_popup(request):
